Refactored file and added screen kt file
