+++ app/src/test/java/com/example/strength4mom/ui/test/ExoViewModelTest.kt
@@ -1,0 +1,13 @@
+package com.example.strength4mom.ui.test
+
+import com.example.strength4mom.ui.theme.exo.ExoViewModel
+import org.junit.Test
+
+class ExoViewModelTest {
+    private val viewModel = ExoViewModel()
+
+    @Test
+    fun testSomethingFromViewModel() {
+
+    }
+}
